+++ src/main/resources/banner.txt
@@ -1,0 +1,6 @@
+ _______         .__       .__   ___.                                            .__              
+ \      \   ____ |__| ____ |  |__\_ |__   ___________    ______ ______________  _|__| ____  ____  
+ /   |   \_/ __ \|  |/ ___\|  |  \| __ \ /  _ \_  __ \  /  ___// __ \_  __ \  \/ /  |/ ___\/ __ \ 
+/    |    \  ___/|  / /_/  >   Y  \ \_\ (  <_> )  | \/  \___ \\  ___/|  | \/\   /|  \  \__\  ___/ 
+\____|__  /\___  >__\___  /|___|  /___  /\____/|__|    /____  >\___  >__|    \_/ |__|\___  >___  >
+        \/     \/  /_____/      \/    \/                    \/     \/                    \/    \/
