+++ src/main/resources/db/migration/V1.0.6__grants.sql
@@ -1,0 +1,3 @@
+GRANT ALL PRIVILEGES ON SCHEMA nb_server TO nb_server;
+GRANT ALL PRIVILEGES ON ALL TABLES IN schema nb_server TO nb_server;
+GRANT ALL PRIVILEGES ON ALL SEQUENCES IN schema nb_server TO nb_server;
