+++ src/test/resources/application.properties
@@ -12,10 +12,9 @@ spring.datasource.password=pKAhppfVvu8z9YY7m2AECjTGpza0onZPeNO7p98r
 spring.datasource.schema=nb_server
 
 #Hibernate Configuration
-#todo remove before prod
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.hibernate.naming.implicit-strategy=org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy
-spring.jpa.show-sql=true
+spring.jpa.show-sql=false
 
 flyway.enabled=true
 flyway.baseline-on-migrate=true
